Refactor: User 엔티티와 인증용 UserDetails 클래스 분리
- 기존에 User 엔티티 자체에 UserDetails 구현
- User 를 순수 도메인 엔티티로 변경 후, UserDetails 를 구현한 AuthUser 클래스 추가
